Added basic Hello world calls. Fixed spring structure problems.
