--- Server/src/main/java/com/livros_livres/Private_Controllers/Colaboradores/FuncionarioController.java
@@ -1,10 +1,0 @@
-package Server.Private_Controllers.Colaboradores;
-import Server.Registers.Usuarios.Funcionario;
-
-public class FuncionarioController {
-
-	public Funcionario getFuncionarioById(int id) {
-		return null;
-	}
-
-}
