--- Server/src/main/java/com/livros_livres/Private_Controllers/Colaboradores/UsuarioController.java
@@ -1,10 +1,0 @@
-package Server.Private_Controllers.Colaboradores;
-import Server.Registers.Usuarios.Usuario;
-
-public class UsuarioController {
-
-	public Usuario getUsuarioById(int id) {
-		return null;
-	}
-
-}
