--- Server/src/main/java/com/livros_livres/Public_Controlllers/Clientes/ClienteController.java
@@ -1,8 +1,0 @@
-import Server.Registers.Usuarios.Cliente;
-
-public class ClienteController {
-
-	public Cliente getClienteById(int id) {
-		return null;
-	}
-}
