--- Server/src/main/java/com/livros_livres/Registers/Livros/Livro.java
@@ -1,8 +1,0 @@
-package Server.Registers.Livros;
-
-public class Livro{
-
-    public Livro(){
-
-    }
-}
