+++ Server/src/main/java/com/livros_livres/Server/ServerApplication.java
@@ -2,8 +2,9 @@ package com.livros_livres.Server;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
+import org.springframework.boot.autoconfigure.jdbc.DataSourceAutoConfiguration;
 
-@SpringBootApplication
+@SpringBootApplication(exclude = {DataSourceAutoConfiguration.class })
 public class ServerApplication {
 
 	public static void main(String[] args) {
