+++ Server/src/main/java/com/livros_livres/Server/private_controllers/colaboradores/UsuarioController.java
@@ -1,0 +1,11 @@
+package com.livros_livres.Server.private_controllers.colaboradores;
+
+import com.livros_livres.Server.registers.usuarios.*;
+
+public class UsuarioController {
+
+	public Usuario getUsuarioById(int id) {
+		return null;
+	}
+
+}
