+++ Server/src/main/java/com/livros_livres/Server/registers/livros/Livro.java
@@ -1,0 +1,8 @@
+package com.livros_livres.Server.registers.livros;
+
+public class Livro{
+
+    public Livro(){
+
+    }
+}
