+++ Server/src/main/java/com/livros_livres/Server/registers/usuarios/Cliente.java
@@ -1,4 +1,4 @@
-package Server.Registers.Usuarios;
+package com.livros_livres.Server.registers.usuarios;
 
 public class Cliente { // Extends Usuario
